refactor: move methods in `mem` module
To use these functions in `acpi` module.
