+++ kernel/src/mem/accessor.rs
@@ -1,15 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::mem::{
-    allocator::{phys::FRAME_MANAGER, virt},
-    paging::pml4::PML4,
-};
-use core::{convert::TryFrom, marker::PhantomData, mem, ptr};
+use core::{marker::PhantomData, mem, ptr};
 use os_units::Bytes;
-use x86_64::{
-    structures::paging::{Mapper, Page, PageSize, PageTableFlags, PhysFrame, Size4KiB},
-    PhysAddr, VirtAddr,
-};
+use x86_64::{PhysAddr, VirtAddr};
 
 pub struct Accessor<T: ?Sized> {
     virt: VirtAddr,
@@ -22,7 +15,7 @@ impl<T> Accessor<T> {
     pub unsafe fn new(phys_base: PhysAddr, offset: Bytes) -> Self {
         let phys_base = phys_base + offset.as_usize();
         let bytes = Bytes::new(mem::size_of::<T>());
-        let virt = Self::map_pages(phys_base, bytes);
+        let virt = super::map_pages(phys_base, bytes);
 
         Self {
             virt,
@@ -55,7 +48,7 @@ impl<T> Accessor<[T]> {
     pub fn new_slice(phys_base: PhysAddr, offset: Bytes, len: usize) -> Self {
         let phys_base = phys_base + offset.as_usize();
         let bytes = Bytes::new(mem::size_of::<T>() * len);
-        let virt = Self::map_pages(phys_base, bytes);
+        let virt = super::map_pages(phys_base, bytes);
 
         Self {
             virt,
@@ -97,7 +90,7 @@ impl<T: ?Sized> Accessor<T> {
     /// object.
     pub unsafe fn new_from_bytes(phys_base: PhysAddr, offset: Bytes, bytes: Bytes) -> Self {
         let phys_base = phys_base + offset.as_usize();
-        let virt = Self::map_pages(phys_base, bytes);
+        let virt = super::map_pages(phys_base, bytes);
 
         Self {
             virt,
@@ -109,53 +102,9 @@ impl<T: ?Sized> Accessor<T> {
     pub fn virt_addr(&self) -> VirtAddr {
         self.virt
     }
-
-    fn map_pages(start: PhysAddr, object_size: Bytes) -> VirtAddr {
-        let start_frame_addr = start.align_down(Size4KiB::SIZE);
-        let end_frame_addr = (start + object_size.as_usize()).align_down(Size4KiB::SIZE);
-
-        let num_pages = Bytes::new(usize::try_from(end_frame_addr - start_frame_addr).unwrap() + 1)
-            .as_num_of_pages::<Size4KiB>();
-
-        let virt = virt::search_free_addr(num_pages)
-            .expect("OOM during creating a new accessor to a register.");
-
-        for i in 0..num_pages.as_usize() {
-            let page = Page::<Size4KiB>::containing_address(virt + Size4KiB::SIZE * i as u64);
-            let frame = PhysFrame::containing_address(start_frame_addr + Size4KiB::SIZE * i as u64);
-            let flag = PageTableFlags::PRESENT;
-
-            unsafe {
-                PML4.lock()
-                    .map_to(page, frame, flag, &mut *FRAME_MANAGER.lock())
-                    .unwrap()
-                    .flush()
-            }
-        }
-
-        let page_offset = start.as_u64() % Size4KiB::SIZE;
-
-        virt + page_offset
-    }
-
-    fn unmap_pages(start: VirtAddr, object_size: Bytes) {
-        let start_frame_addr = start.align_down(Size4KiB::SIZE);
-        let end_frame_addr = (start + object_size.as_usize()).align_down(Size4KiB::SIZE);
-
-        let num_pages = Bytes::new(usize::try_from(end_frame_addr - start_frame_addr).unwrap())
-            .as_num_of_pages::<Size4KiB>();
-
-        for i in 0..num_pages.as_usize() {
-            let page =
-                Page::<Size4KiB>::containing_address(start_frame_addr + Size4KiB::SIZE * i as u64);
-
-            let (_, flush) = PML4.lock().unmap(page).unwrap();
-            flush.flush();
-        }
-    }
 }
 impl<T: ?Sized> Drop for Accessor<T> {
     fn drop(&mut self) {
-        Self::unmap_pages(self.virt, self.bytes)
+        super::unmap_pages(self.virt, self.bytes);
     }
 }
